Deprecation reason returned as empty string triggers "deprecated" state (#3)
GraphQL clients like GraphiQL, Altair, etc expecting value to be
null for active objects. So even with `isDeprecated` set to false,
the object will be considered as deprecated if `deprecationReason`
is just an empty string.

The spec states, that `deprecatedReason` must have string type.
However, in some languages, nullable values are accepted, when in
Golang we have to be explicit if we allow nil values.

I noticed that some GraphQL APIs, such as Github, returns "null"
value with `"isDeprecated": false` and a string when it set to true.

It is unclear, where the bug is, but I don't think it is possible
to fix it for all possible clients and there is a good chance that
old clients will still be showing fields as deprecated.

This commit updates the type from `string` to `*string` for
`DeprecatedReason` attribute, so introspection will return "null"
instead of empty string.